Initalisation fix + Piezo output
After taking out and putting the SD Card back in to the reader, the piezo, the sonar's pins and the IR remote's codes are being logged in the serial monitor. The main thing that is customizable with the code right now (while I wait for Ryan Cather to get back to me with a fix for the IR Remote) is the tone system I can use with the piezo buzzer. An idea that I can use with this is implementing a song library with the IR remote to play different songs on the Arduino piezo buzzer (keep the servo motor the same for volume adjustment (not implemented yet but keep it as is)).